requirements.txt
astroid==2.11.7
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
dill==0.3.5.1
dnspython==2.2.1
email-validator==1.2.1
Flask==2.2.1
Flask-Login==0.6.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
mysql-connector==2.2.9
mysql-connector-python==8.0.30
mysql-connector-python-rf==2.2.2
platformdirs==2.5.2
protobuf==3.20.1
psycopg2==2.9.3
pycparser==2.21
pylint==2.14.5
pylint-flask==0.6
pylint-plugin-utils==0.7
PyMySQL==1.0.2
SQLAlchemy==1.4.40
tomli==2.0.1
tomlkit==0.11.4
Werkzeug==2.2.1
wrapt==1.14.1
WTForms==3.0.1
